+++ apps/film-and-tv-ratings/src/pages/TvShow.tsx
@@ -80,9 +80,10 @@ export function TvShow({ itemId }: { itemId: string }) {
   });
 
   const handleDelete = async (id: string) => {
-    deleteReview({
-      id,
-    });
+    window.confirm('Are you sure you want to delete this review?') &&
+      deleteReview({
+        id,
+      });
   };
 
   const handleEdit = async (id: string) => {
